feat: make connect uploads high priority (#557)
Today, Firehose uploads and user-requested uploads from Connect are treated the same. This is not ideal behavior if the user wants to upload routes immediately for a bug report and the queue is full of uploads for Firehose. The workaround is to clear the queue and retry the upload from Connect.

This PR marks uploads requested by `connect` as highest priority.

Related:
- https://github.com/commaai/openpilot/issues/34836
- https://github.com/commaai/openpilot/pull/34856

Co-authored-by: Cameron Clough <cameronjclough@gmail.com>